Add the ability to avoid known openings.
The GTP command is `lz-analyze avoid-opening <move_number> <directory>`.
It will read all *.sgf files in <directory> and add the first
<move_number> moves of each game to a list of known openings.

Then LZ will only play openings that do not occur in this list.
How it works:

If all <move_number>-1 moves are from a known opening, then the next
move will be avoided as if it were an illegal move.  So if <move_number>
is even, White will need to find a different move, and if it is odd,
Black needs to find a different move.  Of course, LZ already knows about
this at the beginning of the game, so it may choose to play new moves
earlier in order to avoid getting into the variation where the forbidden
move would come up.

This is based on the avoidmove feature from my PR at gcp#1949
I have not added this avoid-opening feature to the PR (yet?).

The GTP command "lz-analyze b avoid-opening <moves> max=<count> <dirpath>"
specifies that only those openings will be avoided that occur more than
<count> times in the SGF files in the directory.
If one opening occurs with two different symmetries, it will be counted
as two occurencies.

The command-line option "--analyze-tags" sets the tags globally, so that
they can be used with the genmove command instead of the "lz-analyze"
command.